+++ joytv/src/com/joyplus/tv/ShowXiangqingTv.java
@@ -1001,6 +1001,8 @@ public class ShowXiangqingTv extends Activity implements View.OnClickListener,
 	
 	public void cancelshoucangResult(String url, JSONObject json, AjaxStatus status){
 		
+		//闹钟开启的情况下，取消收藏删除数据库中相关数据
+		if(StatisticsUtils.is48TimeClock(getApplicationContext()))
 		StatisticsUtils.deleteData4ProId(getApplicationContext(), 
 				StatisticsUtils.getCurrentUserId(getApplicationContext()), prod_id);
